Add a tutorial on creating a custom element.
This guide is a copy of the KKD tutorial on creating a custom element - Markdown editor, only adjusted for the KK official Markdown editor. Language and style changes also applied. The code samples in the guide have been tested to work as a CE when combined into one HTML file.